Moved all components related to header to separate folder
