+++ src/components/header/index.ts
@@ -1,0 +1,2 @@
+export * from './header';
+export * from './actions';
