+++ src/components/index.ts
@@ -1,3 +1,3 @@
 export * from './error-fallback';
-export * from './header';
+export * from './header/header';
 export * from './navbar';
